Bug 1689620 - PerfSheriffBot reports outcome
* resolve conflicts
* provide schema migration for extended backfill records
* delegate content creation
* stabilize or temporary silence part of the tests
* group PerfSheriffBot' s components under single package
* rename PerfSheriffBot' script
* fix platform based limit
* extract Notify service
* abstract away perf sheriff bot' s instantiation
* refactor email writting
* refactor delete notification email
* fix incorrect exception raise
* provide proper subject & content for backfill report email
* merge email modules into one
* provide NullObject for email notifier
* fix tests from BigQuery integration
* fix notification mocks in tests
* build push range link
* raise appropriate exceptions from email composition
* improve Linux perf alert fixture
* handle json error exception when writing email
* exclude abstract methods from coverage
* fix broken PerfSheriffBot test
* ensure TaskclusterModelImpl cannot be instantiated on production
* basic unit tests for email notifications related to expiring signatures
* ensure happy path testing for BackfillNotificationWriter via PerfSheriffBot
* provide coverage for TaskclusterModelImpl' s constructor
* address some PyCharm inspection warnings
* address review requests
* better group tests for deletion email writer
* test email integration in PerfSheriffBot
* provide email writer via dependency injection to ease testability
* turn some private methods into staticmethods
