+++ tests/model/cycle_data/test_perfherder_cycling.py
@@ -4,13 +4,12 @@ from unittest.mock import MagicMock
 from unittest.mock import patch
 
 import pytest
-import taskcluster
 from django.core.management import call_command
 from django.db import connection, IntegrityError
 
 from treeherder.model.data_cycling import MaxRuntime
 from treeherder.model.data_cycling import PerfherderCycler
-from treeherder.model.data_cycling import PublicSignatureRemover, EmailService
+from treeherder.model.data_cycling import PublicSignatureRemover
 from treeherder.model.data_cycling.removal_strategies import (
     MainRemovalStrategy,
     TryDataRemoval,
@@ -43,7 +42,7 @@ def test_cycle_performance_data(
     repository_name,
     push_stored,
     test_perf_signature,
-    mock_taskcluster_notify,
+    taskcluster_notify_mock,
 ):
     test_repository.name = repository_name
     test_repository.save()
@@ -81,7 +80,7 @@ def test_cycle_performance_data(
         value=1.0,
     )
 
-    # the performance datum that which we're targetting
+    # the performance datum that which we're targeting
     PerformanceDatum.objects.create(
         id=2,
         repository=test_repository,
@@ -104,7 +103,7 @@ def test_cycle_performance_data(
     ]
 
 
-def test_performance_signatures_are_deleted(test_perf_signature, mock_taskcluster_notify):
+def test_performance_signatures_are_deleted(test_perf_signature, taskcluster_notify_mock):
     cycler = PerfherderCycler(chunk_size=100, sleep_time=0)
     expired_timestamp = cycler.max_timestamp
 
@@ -148,7 +147,7 @@ def test_try_data_removal(
     try_push_stored,
     test_perf_signature,
     test_perf_signature_2,
-    mock_taskcluster_notify,
+    taskcluster_notify_mock,
 ):
     total_removals = 3
     test_perf_signature.repository = try_repository
@@ -206,7 +205,7 @@ def test_irrelevant_repos_data_removal(
     repository_name,
     push_stored,
     test_perf_signature,
-    mock_taskcluster_notify,
+    taskcluster_notify_mock,
 ):
     # test_repository is considered irrelevant repositories
 
@@ -265,7 +264,7 @@ def test_signature_remover(
     test_perf_signature,
     test_perf_signature_2,
     test_perf_data,
-    mock_taskcluster_notify,
+    taskcluster_notify_mock,
     mock_tc_prod_credentials,
 ):
     cycler = PerfherderCycler(chunk_size=100, sleep_time=0)
@@ -277,57 +276,23 @@ def test_signature_remover(
 
     call_command('cycle_data', 'from:perfherder')
 
-    assert taskcluster.Notify().ping.called_once
-    assert taskcluster.Notify().email.call_count == 1
+    assert taskcluster_notify_mock.email.call_count == 1
     assert len(PerformanceSignature.objects.all()) == 1
     assert PerformanceSignature.objects.first() == test_perf_signature
-
-
-def test_email_content(test_perf_signature):
-    email_service = EmailService(address='test_email@test.com')
-    email_service.add_signature_to_content(test_perf_signature)
-    expected_result = email_service.TABLE_DESCRIPTION + email_service.TABLE_HEADERS
-    expected_result += (
-        """| {repository} | {framework} | {platform} | {suite} | {application} |""".format(
-            repository=test_perf_signature.repository.name,
-            framework=test_perf_signature.framework.name,
-            platform=test_perf_signature.platform.platform,
-            suite=test_perf_signature.suite,
-            application=test_perf_signature.application,
-        )
-    )
-    expected_result += '\n'
-    assert expected_result == email_service.content
-
-
-def test_signature_remover_when_notify_service_is_down(
-    test_perf_signature,
-    test_perf_signature_2,
-    test_perf_data,
-    mock_taskcluster_notify,
-    mock_tc_prod_credentials,
-):
-    taskcluster.Notify().ping.side_effect = Exception('Email Service is down.')
-
-    call_command('cycle_data', 'from:perfherder')
-
-    assert taskcluster.Notify().ping.called_once
-    assert not taskcluster.Notify().email.called
-    assert len(PerformanceSignature.objects.all()) == 2
 
 
 @pytest.mark.parametrize('total_signatures', [3, 4, 8, 10])
 def test_total_emails_sent(
     test_perf_signature, try_repository, total_signatures, mock_tc_prod_credentials
 ):
-    tc_model = MagicMock()
+    notify_client_mock = MagicMock()
     timer = MaxRuntime()
     timer.start_timer()
     total_rows = 2
     total_emails = 4
     signatures_remover = PublicSignatureRemover(
         timer=timer,
-        taskcluster_model=tc_model,
+        notify_client=notify_client_mock,
         max_rows_allowed=total_rows,
         max_emails_allowed=total_emails,
     )
@@ -371,21 +336,21 @@ def test_total_emails_sent(
     signatures = PerformanceSignature.objects.filter(last_updated__lte=datetime.now())
     signatures_remover.remove_in_chunks(signatures)
 
-    assert tc_model.notify.email.call_count == expected_call_count
+    assert notify_client_mock.email.call_count == expected_call_count
     assert not PerformanceSignature.objects.filter(repository__name='try').exists()
 
 
 def test_remove_try_signatures_without_data(
     test_perf_signature, test_perf_data, try_repository, mock_tc_prod_credentials
 ):
-    tc_model = MagicMock()
+    notify_client_mock = MagicMock()
     timer = MaxRuntime()
     timer.start_timer()
     total_rows = 2
     total_emails = 2
     signatures_remover = PublicSignatureRemover(
         timer=timer,
-        taskcluster_model=tc_model,
+        notify_client=notify_client_mock,
         max_rows_allowed=total_rows,
         max_emails_allowed=total_emails,
     )
@@ -418,7 +383,7 @@ def test_remove_try_signatures_without_data(
     assert PerformanceSignature.objects.filter(id=signature_with_perf_data.id).exists()
 
 
-def test_performance_cycler_quit_indicator(mock_taskcluster_notify):
+def test_performance_cycler_quit_indicator(taskcluster_notify_mock):
     ten_minutes_ago = datetime.now() - timedelta(minutes=10)
     one_second = timedelta(seconds=1)
 
@@ -463,7 +428,7 @@ def empty_alert_summary(
     ],
 )
 def test_summary_without_any_kind_of_alerts_is_deleted(
-    expired_time, empty_alert_summary, mock_taskcluster_notify
+    expired_time, empty_alert_summary, taskcluster_notify_mock
 ):
     empty_alert_summary.created = expired_time
     empty_alert_summary.save()
@@ -486,7 +451,7 @@ def test_summary_without_any_kind_of_alerts_is_deleted(
     ],
 )
 def test_summary_without_any_kind_of_alerts_isnt_deleted(
-    recently, empty_alert_summary, mock_taskcluster_notify
+    recently, empty_alert_summary, taskcluster_notify_mock
 ):
     empty_alert_summary.created = recently
     empty_alert_summary.save()
@@ -519,7 +484,7 @@ def test_summary_with_alerts_isnt_deleted(
     test_perf_alert,
     test_perf_alert_2,
     test_perf_data,
-    mock_taskcluster_notify,
+    taskcluster_notify_mock,
 ):
     empty_alert_summary.created = creation_time
     empty_alert_summary.save()
