+++ tests/perfalert/test_auto_perf_sheriffing/test_backfill_reports/test_alerts_picker.py
@@ -1,91 +1,9 @@
 from collections import Counter
-from copy import deepcopy
 from unittest.mock import Mock
 
 import pytest
 
-from treeherder.perf.alerts import AlertsPicker
-from treeherder.perf.models import PerformanceAlert
-
-
-@pytest.fixture
-def test_many_various_alerts():
-    alerts = [Mock(spec=PerformanceAlert) for i in range(10)]
-    platforms = (
-        'windows10-64-shippable',
-        'windows10-64-shippable',
-        'windows7-32-shippable',
-        'windows7-32-shippable',
-        'linux64-shippable-qr',
-        'linux64-shippable-qr',
-        'osx-10-10-shippable',
-        'osx-10-10-shippable',
-        'android-hw-pix-7-1-android-aarch64',
-        'android-hw-pix-7-1-android-aarch64',
-    )
-
-    reversed_magnitudes = list(reversed(range(len(alerts))))
-    toggle = True
-    for idx, alert in enumerate(alerts):
-        alert.is_regression = toggle
-        alert.series_signature.platform.platform = platforms[idx]
-        alert.amount_pct = reversed_magnitudes[idx]
-        toggle = not toggle
-    return alerts
-
-
-@pytest.fixture
-def test_few_various_alerts():
-    alerts = [Mock(spec=PerformanceAlert) for i in range(2)]
-    platforms = ('windows7-32-shippable', 'linux64-shippable-qr')
-    reversed_magnitudes = list(reversed(range(len(alerts))))
-    toggle = True
-    for idx, alert in enumerate(alerts):
-        alert.series_signature.platform.platform = platforms[idx]
-        alert.is_regression = toggle
-        alert.amount_pct = reversed_magnitudes[idx]
-        toggle = not toggle
-    return alerts
-
-
-@pytest.fixture
-def test_few_regressions():
-    alerts = [Mock(spec=PerformanceAlert) for i in range(5)]
-    platforms = (
-        'windows10-64-shippable',
-        'windows7-32-shippable',
-        'linux64-shippable-qr',
-        'osx-10-10-shippable',
-        'android-hw-pix-7-1-android-aarch64',
-    )
-    reversed_magnitudes = list(reversed(range(len(alerts))))
-    for idx, alert in enumerate(alerts):
-        alert.series_signature.platform.platform = platforms[idx]
-        alert.is_regression = True
-        alert.amount_pct = reversed_magnitudes[idx]
-    return alerts
-
-
-@pytest.fixture
-def test_few_improvements(test_few_regressions):
-    alerts = deepcopy(test_few_regressions)
-    for alert in alerts:
-        alert.is_regression = False
-    return alerts
-
-
-@pytest.fixture
-def test_bad_platform_names():
-    alerts = [Mock(spec=PerformanceAlert) for i in range(4)]
-    platforms = (
-        'rfvrtgb',
-        '4.0',
-        '54dcwec58',
-        '8y6 t g',
-    )
-    for idx, alert in enumerate(alerts):
-        alert.series_signature.platform.platform = platforms[idx]
-    return alerts
+from treeherder.perf.auto_perf_sheriffing.backfill_reports import AlertsPicker
 
 
 def test_init():
@@ -141,9 +59,9 @@ def test_extract_important_alerts(
     number_of = count_alert_types(important_alerts)
     assert number_of[True] == 4
     assert number_of[False] == 1
-    for idx, alert in enumerate(important_alerts):
-        assert alert.series_signature.platform.platform == expected_platforms_order[idx]
-        assert alert.amount_pct == expected_magnitudes_order[idx]
+    for idx, imp_alert in enumerate(important_alerts):
+        assert imp_alert.series_signature.platform.platform == expected_platforms_order[idx]
+        assert imp_alert.amount_pct == expected_magnitudes_order[idx]
 
 
 def test_ensure_alerts_variety(
@@ -194,7 +112,7 @@ def test_ensure_alerts_variety(
 
 
 @pytest.mark.parametrize(
-    ('max_alerts, expected_alerts_platforms'),
+    ('max_alerts, expected_alerts_platforms'),  # noqa
     [
         (5, ('windows10', 'windows7', 'linux', 'osx', 'android')),
         (8, ('windows10', 'windows7', 'linux', 'osx', 'android', 'windows10', 'windows7', 'linux')),
@@ -225,7 +143,7 @@ def test_os_relevance():
     assert 1 == picker._os_relevance('android')
 
     with pytest.raises(ValueError):
-        picker._os_relevance('some wierd OS')
+        picker._os_relevance('some weird OS')
 
 
 def test_has_relevant_platform(test_many_various_alerts, test_bad_platform_names):
@@ -282,6 +200,6 @@ def test_multi_criterion_sort(test_many_various_alerts):
     number_of = count_alert_types(ordered_alerts)
     assert number_of[True] == 5
     assert number_of[False] == 5
-    for idx, alert in enumerate(ordered_alerts):
-        assert alert.series_signature.platform.platform == expected_platforms_order[idx]
-        assert alert.amount_pct == expected_magnitudes_order[idx]
+    for idx, ord_alert in enumerate(ordered_alerts):
+        assert ord_alert.series_signature.platform.platform == expected_platforms_order[idx]
+        assert ord_alert.amount_pct == expected_magnitudes_order[idx]
