+++ treeherder/model/data_cycling/__init__.py
@@ -1,4 +1,3 @@
 from .cyclers import TreeherderCycler, PerfherderCycler, TREEHERDER, PERFHERDER  # noqa
 from .max_runtime import MaxRuntime  # noqa
 from .signature_remover import PublicSignatureRemover  # noqa
-from .email_service import EmailService  # noqa
