+++ treeherder/model/data_cycling/cyclers.py
@@ -7,12 +7,11 @@ from django.db import OperationalError, connection
 from django.db.backends.utils import CursorWrapper
 from django.db.models import Count
 
-from treeherder.config import settings
 from treeherder.model.data_cycling.removal_strategies import RemovalStrategy
 from treeherder.model.models import Job, JobType, JobGroup, Machine
 from treeherder.perf.exceptions import NoDataCyclingAtAll, MaxRuntimeExceeded
 from treeherder.perf.models import PerformanceSignature, PerformanceAlertSummary
-from treeherder.services.taskcluster import TaskclusterModel, DEFAULT_ROOT_URL as root_url
+from treeherder.services import taskcluster
 from .max_runtime import MaxRuntime
 from .signature_remover import PublicSignatureRemover
 from .utils import has_valid_explicit_days
@@ -134,10 +133,8 @@ class PerfherderCycler(DataCycler):
         # remove any signatures which are
         # no longer associated with a job
         signatures = PerformanceSignature.objects.filter(last_updated__lte=self.max_timestamp)
-        tc_model = TaskclusterModel(
-            root_url, client_id=settings.NOTIFY_CLIENT_ID, access_token=settings.NOTIFY_ACCESS_TOKEN
-        )
-        signatures_remover = PublicSignatureRemover(timer=self.timer, taskcluster_model=tc_model)
+        notify_client = taskcluster.notify_client_factory()
+        signatures_remover = PublicSignatureRemover(timer=self.timer, notify_client=notify_client)
         signatures_remover.remove_in_chunks(signatures)
 
         # remove empty alert summaries
