+++ treeherder/perf/models.py
@@ -1,6 +1,5 @@
 import datetime
 import json
-
 from typing import List
 
 from django.contrib.auth.models import User
@@ -9,7 +8,14 @@ from django.core.validators import MinLengthValidator
 from django.db import models
 from django.utils.timezone import now as django_now
 
-from treeherder.model.models import Job, MachinePlatform, OptionCollection, Push, Repository
+from treeherder.model.models import (
+    Job,
+    MachinePlatform,
+    OptionCollection,
+    Push,
+    Repository,
+    JobType,
+)
 from treeherder.utils import default_serializer
 
 SIGNATURE_HASH_LENGTH = 40
@@ -21,6 +27,10 @@ class PerformanceFramework(models.Model):
 
     class Meta:
         db_table = 'performance_framework'
+
+    @classmethod
+    def fetch_all_names(cls) -> List[str]:
+        return cls.objects.values_list('name', flat=True)
 
     def __str__(self):
         return self.name
@@ -494,6 +504,7 @@ class PerformanceBugTemplate(models.Model):
         return '{} bug template'.format(self.framework.name)
 
 
+# TODO: we actually need this name for the PerfSheriffBot' s hourly report
 class BackfillReport(models.Model):
     """
     Groups & stores all context required to retrigger/backfill
@@ -557,7 +568,13 @@ class BackfillRecord(models.Model):
     )
 
     status = models.IntegerField(choices=STATUSES, default=PRELIMINARY)
+
+    # Backfill outcome
     log_details = models.TextField()  # JSON expected, not supported by Django
+    job_type = models.ForeignKey(
+        JobType, null=True, on_delete=models.SET_NULL, related_name='backfill_records'
+    )
+    total_backfills_triggered = models.IntegerField(default=0)
 
     @property
     def id(self):
